Various improvements

+++ appmemdumper/memdump.py
@@ -6,6 +6,10 @@ import shlex
 import shutil
 import StringIO
 import sys
+from copy import deepcopy
+from os.path import abspath, dirname, exists, expanduser, isdir, isfile, join
+from subprocess import check_output, CalledProcessError, PIPE, Popen
+
 import volatility.addrspace as addrspace
 import volatility.conf as conf
 import volatility.commands as commands
@@ -14,9 +18,7 @@ import volatility.debug as debug
 import volatility.obj as obj
 import volatility.registry as registry
 import volatility.utils as utils
-from copy import deepcopy
-from os.path import abspath, dirname, exists, isdir, isfile
-from subprocess import check_output, CalledProcessError, PIPE, Popen
+
 from .dumpers import *
 
 try:
@@ -26,12 +28,12 @@ except OSError:
     sys.exit(1)
 
 
-__all__ = ["DUMPERS", "VolatilityMemDump"]
+__all__ = ["APPDUMPERS", "SYSDUMPERS", "VolatilityMemDump"]
 CMD = 'volatility {opt} {cmd}'
+logger = logging.getLogger("main")
 
 
 is_memdump = lambda x: isinstance(x, VolatilityMemDump)
-logger = logging.getLogger("main")
 
 
 class VolatilityAppDumper(object):
@@ -64,20 +66,19 @@ class VolatilityAppDumper(object):
         # second, find the right application dumper for the given process name
         #  and attach the dumper to the memory dump if a dumper is found and if
         #  the related application is handled
-        # if no process name was given and there exists a general-purpose dumper
-        #  (that has no list of process names), create a dumper
-        for cls in DUMPERS:
+        # if no process name was given and there exists a system dumper (that
+        #  has no list of process names), create a dumper
+        for cls in APPDUMPERS + SYSDUMPERS:
             pnames = globals()[cls].procnames
             if pnames is not None:
                 pnames = [p.lower() for p in pnames]
             if (pname is None and pnames is None) or \
                (pname is not None and pnames is not None and pname in pnames):
-                if cls in dump._selected_apps:
+                if cls in dump._selected_apps or cls in dump._selected_syst:
                     if pname is not None:
                         dump._stats[0].append(pname)  # found
-                    name = pname.split('.', 1)[0] if pname is not None else cls
-                    self.dumper = globals()[cls](dump, name, pids)
-                    dump.apps[cls] = self.dumper
+                    self.dumper = globals()[cls](dump, pids)
+                    dump.dumpers[cls] = self.dumper
                 else:
                     if pname is not None:
                         dump._stats[1].append(pname)  # not selected
@@ -123,14 +124,15 @@ class VolatilityMemDump(object):
                                ["Suggested Profile(s)"].split(", "),
     }
 
-    def __init__(self, dump, selected_apps, out_dir="files", plugins_dir=None,
+    def __init__(self, dump, apps, syst, out_dir="files", plugins_dir=None,
                  from_cache=True):
         """
         Determines the profile, retrieves the list of processes and creates the
          list of application dumpers.
 
         :param dump: Volatility memory dump filename
-        :param selected_apps: list of the application names to be handled
+        :param apps: list of the application dumper classes to be handled
+        :param syst: list of the system dumper classes to be handled
         :param out_dir: output directory for retrieved resources
         :param plugins_dir: Volatility custom plugins directory
         :param from_cache: boolean indicating if the profile must be retrieved
@@ -139,17 +141,25 @@ class VolatilityMemDump(object):
         short = dump
         dump = abspath(dump)
         assert isfile(dump), "{} is not a dump file".format(dump)
-        assert all(x in DUMPERS for x in selected_apps), "Unknown dumper(s)"
+        assert all(x in APPDUMPERS for x in apps), "Unknown application dumper(s)"
+        assert all(x in SYSDUMPERS for x in syst), "Unknown system dumper(s)"
         assert plugins_dir is None or isdir(plugins_dir), "Bad plugins dir"
         assert isinstance(from_cache, bool)
-        self._selected_apps = selected_apps
+        self._cache = {}
+        self._selected_apps = apps
+        self._selected_syst = syst
+        if len(self._selected_apps) == 0 and len(self._selected_syst) == 0:
+            logger.warning("No dumper selected")
+            sys.exit(0)
         logger.debug("Setting output directory to '{}'...".format(out_dir))
         self.out_dir = abspath(out_dir)
+        self._cachefile = join(self.out_dir, ".cache")
         # initialize dump opening
         registry.PluginImporter()
         self.__is_profile_tested = False
         self.config = conf.ConfObject()
         if plugins_dir is not None:
+            plugins_dir = expanduser(plugins_dir)
             logger.debug("Setting plugins directory to '{}'..."
                          .format(plugins_dir))
             self.config.plugins = abspath(plugins_dir)
@@ -170,17 +180,17 @@ class VolatilityMemDump(object):
         :param from_cache: load the profile from the cache file if it exists
         :return: boolean indicating if the profile is tested as correct
         """
+        cf = self._cachefile
         assert isinstance(from_cache, bool)
         logger.info("Getting profile...")
         # get available profiles and commands
         available_profiles = registry.get_plugin_classes(obj.Profile)
         compatible_profiles = []
         # load profile from cache or use the 'imageinfo' command to find it
-        tmp = "/tmp/appmemdumper/profile"
         if from_cache:
-            if isfile(tmp):
-                with open(tmp) as f:
-                    self.config.PROFILE = f.read().strip()
+            if isfile(cf):
+                with open(cf) as f:
+                    self.config.PROFILE = f.readlines()[0].strip()
                 if not self.__get_pslist():
                     self.config.PROFILE = None
                 else:
@@ -205,11 +215,7 @@ class VolatilityMemDump(object):
             logger.error("No suitable profile could be found ; please check "
                          "that your memory dump is supported by Volatility")
             sys.exit(2)
-        # create a cache file with the profile name
-        if not isdir(dirname(tmp)):
-            os.makedirs(dirname(tmp))
-        with open(tmp, 'w') as f:
-            f.write(self.config.PROFILE)
+        self._update_cache(self.config.PROFILE)
         logger.info("> Selected profile: {}".format(self.config.PROFILE))
         return is_profile_tested
 
@@ -226,28 +232,67 @@ class VolatilityMemDump(object):
         logger.info("Getting processes...")
         try:
             self.processes = self.execute('pslist')
+            self.processes.extend(self.execute("psscan"))
         except:
             logger.debug("Wrong profile, re-getting profile...")
             return False
+        # collect PIDs relationships
+        procs = {'0': "Init"}
+        for p in self.processes:
+            procs[str(int(p.UniqueProcessId))] = p.ImageFileName
+        self.ppids = {}
+        for p in self.processes:
+            pid = str(int(p.UniqueProcessId))
+            ppid = str(int(p.InheritedFromUniqueProcessId))
+            try:
+                k = (ppid, procs[ppid])
+            except KeyError:
+                k = (ppid, None)
+            self.ppids.setdefault(k, [])
+            self.ppids[k].append((pid, procs[pid]))
         # parse the list of processes and create objects for each set of
         #  application instances that are handled by the present tool
-        self.apps = {}
+        self.dumpers = {}
         self._stats = [[], [], []]
         procnames = set([str(p.ImageFileName) for p in self.processes])
-        # first, search for general-purpose dumpers
+        # first, search for system dumpers
         VolatilityAppDumper(self)
-        # second, search for specific-purpose dumpers per process name
+        # second, search for application dumpers per process name
         for n in sorted(procnames, key=lambda n: n.lower()):
             VolatilityAppDumper(self, n.lower())
         for i, t in enumerate(["Found", "Not selected", "Not handled"]):
             s = self._stats[i]
             if len(s) > 0:
                 logger.debug("{}: {}".format(t.ljust(12), ", ".join(s)))
-        if len(self.apps) == 0:
+        if len(self.dumpers) == 0:
             logger.warning("No application to be handled")
         return True
+    
+    def _update_cache(self, profile=None, dumper=None):
+        """
+        Update cache file for the current memory dump.
+        
+        :param profile: profile to be written to the cache
+        :param dumper: new dumper to be added to the cache
+        """
+        if profile is None and dumper is None:
+            return
+        _ = self._cachefile
+        if not exists(_):
+            with open(_, 'w') as f:
+                f.write(self.config.PROFILE)
+        with open(_) as f:
+            c = f.read().splitlines()
+        p, d = c[0], c[1:]
+        if profile is not None:
+            c[0] = profile
+        if dumper is not None:
+            d = sorted(list(set(d + [dumper])))
+            c[1:] = d
+        with open(_, 'w') as f:
+            f.write('\n'.join(c))
 
-    def call(self, command, options=None, parser=None, failmode="error"):
+    def call(self, command, options=None, parser=None, failmode="silent"):
         """
         Run Volatility with given command and options using subprocess.
 
@@ -270,6 +315,12 @@ class VolatilityMemDump(object):
         if self.config.plugins is not None and len(self.config.plugins) > 0:
             options.insert(0, "--plugins={}".format(self.config.plugins))
         cmd = CMD.format(cmd=command, opt=" ".join(options))
+        # try to get the result from cache first
+        if cmd in self._cache:
+            logger.debug("> Got result of command '{}' from cache..."
+                         .format(command))
+            logger.debug(">> {}".format(cmd))
+            return self._cache[cmd]
         # run the resulting command with a subprocess
         logger.debug("> Calling command '{}'...".format(command))
         logger.debug(">> {}".format(cmd))
@@ -279,7 +330,7 @@ class VolatilityMemDump(object):
             if failmode == "silent":
                 return
             if "You must specify something to do (try -h)" in err:
-                msg = "Plugin does not exist"
+                msg = "Plugin '{}' not found".format(command)
             else:
                 msg = "({}) {}".format(*map(lambda x: x.strip(),
                                             err.split(":", 2)[1:3]))
@@ -293,28 +344,42 @@ class VolatilityMemDump(object):
         # then parse the output with the given parser and return the result
         if parser is not None:
             out = parser(out)
+        # cache the result
+        self._cache[cmd] = out
         return out
 
-    def dump(self):
+    def dump(self, update=False):
         """
         Dump resources for every known application in the memory dump.
+        
+        :param update: update output directory (meaning that collected pieces of
+                        information should not be recollected using Volatility
+                        commands)
         """
-        if exists(self.out_dir):
+        if not update and exists(self.out_dir):
             shutil.rmtree(self.out_dir) if isdir(self.out_dir) \
                 else os.remove(self.out_dir)
-        os.makedirs(self.out_dir)
+        elif update and isfile(self.out_dir):
+            logger.error("Output directory is not valid")
+            sys.exit(2)
+        if not exists(self.out_dir):
+            os.makedirs(self.out_dir)
         # run each dumper and collect information messages for guidance
+        d = self.dumpers
+        logger.info("Processing {} dumper{}{}..."
+                    .format(len(d), ["", "s"][len(d) > 1],
+                           ["", ", this may take a while"][len(d) > 3]))
         msgs = {}
-        for n, a in self.apps.items():
-            logger.info("Processing dumper '{}'...".format(n))
-            for msg in a._run():
+        for n, a in d.items():
+            logger.debug("Processing dumper '{}'...".format(n))
+            for msg in a._run(update):
                 msgs.setdefault(msg, [])
                 msgs[msg].append(n)
         # finally, warn the collected messages
         for m, n in msgs.items():
             logger.warning("\nThe following applies to collected objects of:\n-"
                            " {}\n\n{}\n".format("\n- ".join(n), m))
-            
+    
     def execute(self, command, nopid=False, text=False, parser=None):
         """
         Run Volatility command through the dedicated Python API.
@@ -330,11 +395,16 @@ class VolatilityMemDump(object):
         assert isinstance(nopid, bool)
         assert isinstance(text, bool)
         assert parser is None or callable(parser)
+        # try to get the result from cache first
+        if command in self._cache:
+            logger.debug("> Got result of command '{}' from cache..."
+                         .format(command))
+            return self._cache[command]
+        # run the command with the API
         logger.debug("> Executing command '{}'...".format(command))
         if nopid:
             pid = self.config.PID
             self.config.PID = None
-        # run the command with the API
         command = self.__commands[command](deepcopy(self.config))
         result = command.calculate()
         # convert the result to text if required
@@ -349,4 +419,6 @@ class VolatilityMemDump(object):
         # then parse the output with the given parser and return the result
         if parser:
             out = parser(out)
+        # cache the result
+        self._cache[command] = out
         return out
